This is the first (working) commit that allows iCal attachments for meeting mails. Hence, the mail client can provide a way to easily pop it in the calender.
This requires the standard (http://www.ietf.org/rfc/rfc2445.txt) which is actually accomplished by use of RiCal (see http://ri-cal.rubyforge.org/, http://ri-cal.rubyforge.org/).
The Gemfile is added to deal with this.

The meeting itself has a start_hour and end_hour which is not defined for any time_zone which apperently is very critical for these calender apps - hence the timeZone assumed is that of local.

This has been tested under 2.2.x - with most times being accurate.
The timezone name and date-time-string computation seems to be a little clumsy which I am planning to clean up in next commit.

Any feedback would be welcome.

Dipan Mehta.